requirements.txt
# Mutalyze Requirements
# Comprehensive Protein Mutation Stability Analysis Platform

# Core Dependencies
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Bioinformatics
biopython>=1.81

# Molecular Dynamics (Optional but recommended)
# Install via conda: conda install -c conda-forge openmm
# openmm>=8.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
py3Dmol>=2.0.0

# Parallel Processing
tqdm>=4.65.0

# Web and Data
requests>=2.31.0

# External Tools (Install separately)
# BLAST+: conda install -c bioconda blast
# MAFFT: conda install -c bioconda mafft

# Development Dependencies (Optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
